Move customize controls into seperate component
